\name{polygon_area}

\alias{polygon_area}

\title{
	Calculates the Area of a Given Polygon
}

\description{
	For internal use only.
}

\usage{
polygon_area(poly)
}

\arguments{
  \item{poly}{
	a matrix with two columns (i.e., x & y) and where the first and last rows are equal.
	}
}

\value{
	the area of the polygon.
}

\author{
	anonymous
}

\seealso{
	\code{\link{rase}}
}