\name{polygon_centroid}

\alias{polygon_centroid}

\title{
	Calculate the Polygon Center
}

\description{
	For internal use only.
}

\usage{
	polygon_centroid(poly)
}

\arguments{
  \item{poly}{
	a matrix with two columns (i.e., x & y) and where the first and last rows are equal.
	}
}

\value{
 	a vector of the polygon centroid coordinates.
}
\author{
	anonymous
}

\seealso{
	\code{\link{rase}}
}

