% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.dnn.R
\name{predict.dnn}
\alias{predict.dnn}
\title{Predict a deep learning model via keras.}
\usage{
\method{predict}{dnn}(object, newdata, ...)
}
\arguments{
\item{object}{a}

\item{newdata}{a}

\item{...}{a}
}
\value{
Predicted response.
\item{ypred}{The predicted response.}
}
\description{
Predict a deep learning model.
}
\references{
Tian, Y. and Feng, Y., 2021(a). RaSE: A variable screening framework via random subspace ensembles. Journal of the American Statistical Association, (just-accepted), pp.1-30.

Tian, Y. and Feng, Y., 2021(b). RaSE: Random subspace ensemble classification. Journal of Machine Learning Research, 22(45), pp.1-93.

Zhu, J. and Feng, Y., 2021. Super RaSE: Super Random Subspace Ensemble Classification. https://www.preprints.org/manuscript/202110.0042

Chen, J. and Chen, Z., 2008. Extended Bayesian information criteria for model selection with large model spaces. Biometrika, 95(3), pp.759-771.

Chen, J. and Chen, Z., 2012. Extended BIC for small-n-large-P sparse GLM. Statistica Sinica, pp.555-574.

Akaike, H., 1973. Information theory and an extension of the maximum likelihood principle. In 2nd International Symposium on Information Theory, 1973 (pp. 267-281). Akademiai Kaido.

Schwarz, G., 1978. Estimating the dimension of a model. The annals of statistics, 6(2), pp.461-464.
}
