% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mRaSE.R
\name{print.mRaSE}
\alias{print.mRaSE}
\title{Print a fitted mRaSE object.}
\usage{
\method{print}{mRaSE}(x, ...)
}
\arguments{
\item{x}{fitted \code{'mRaSE'} model object.}

\item{...}{additional arguments.}
}
\value{
No value is returned.
}
\description{
Similar to the usual print methods, this function summarizes results.
from a fitted \code{'mRaSE'} object.
}
\examples{
set.seed(0, kind = "L'Ecuyer-CMRG")
train.data <- RaModel("classification", 1, n = 100, p = 50)
xtrain <- train.data$x
ytrain <- train.data$y
\dontrun{
# test RaSE classifier with LDA base classifier
fit <- SmultiRase(xtrain, ytrain, B1 = 50, B2 = 50, iteration = 0, cutoff = TRUE,
base = 'lda', cores = 2, ranking = TRUE)

# print the summarized results
print(fit)
}
}
\seealso{
\code{\link{RaSE}}.
}
