% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmultiRaSE.R
\name{SmultiRase}
\alias{SmultiRase}
\title{Construct the random subspace ensemble classifier.}
\usage{
SmultiRase(
  xtrain,
  ytrain,
  xval = NULL,
  yval = NULL,
  B1 = 50,
  B2 = 100,
  D = NULL,
  dist = NULL,
  base = "lda",
  super = list(type = "seperate", base.update = TRUE),
  criterion = "cv",
  ranking = TRUE,
  k = c(3, 5, 7, 9, 11),
  cores = 1,
  seed = NULL,
  iteration = 1,
  cutoff = TRUE,
  cv = 5,
  scale = FALSE,
  C0 = 0.1,
  kl.k = NULL,
  lower.limits = NULL,
  upper.limits = NULL,
  weights = NULL,
  ...
)
}
\description{
\code{RaSE} is a general ensemble classification framework to solve the sparse classification problem. In RaSE algorithm, for each of the B1 weak learners, B2 random subspaces are generated and the optimal one is chosen to train the model on the basis of some criterion.
}
