% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.SmRaSE.R
\name{predict.SmRaSE}
\alias{predict.SmRaSE}
\title{Predict the outcome of new observations based on the estimated super mRaSE classifier (Bi, F., Zhu, J. and Feng, Y., 2022).}
\usage{
\method{predict}{SmRaSE}(object, newx, type = c("vote", "prob", "raw-vote", "raw-prob"), ...)
}
\arguments{
\item{object}{a}

\item{newx}{a}

\item{type}{a}

\item{...}{a}
}
\description{
Predict the outcome of new observations based on the estimated super mRaSE classifier (Bi, F., Zhu, J. and Feng, Y., 2022).
}
\examples{
set.seed(0, kind = "L'Ecuyer-CMRG")
train.data <- RaModel("multi_classification", model.no = 1, n = 100,
p = 50, p0 = rep(1/4,4))
test.data <- RaModel("multi_classification", model.no = 1, n = 100,
p = 50, p0 = rep(1/4,4))
xtrain <- train.data$x
colnames(xtrain) <- paste0("V",1:dim(xtrain)[2])
ytrain <- train.data$y
xtest <- test.data$x
colnames(xtest) <- paste0("V",1:dim(xtest)[2])
ytest <- test.data$y

fit <- SmultiRase(xtrain, ytrain, B1 = 20, B2 = 50, iteration = 0,
base = c('lda','knn'), cores = 1)
ypred <- predict(fit, xtest)
mean(ypred != ytest)
}
